Moved materials to git (#748) (#777)
* Moved the materials and shaders to git

* Updated the version number

* Fixed some compiler warnings and made the move constructors noexcept

Otherwise standard containers don't use the move operations

+++ src/microbe_stage/biomes.cpp
@@ -115,8 +115,8 @@ Biome::Biome(Json::Value value)
         // Add meshes
         Json::Value meshData = chunkData[chunkInternalName]["meshes"];
 
-        for(int i = 0; i < meshData.size(); i++) {
-            chunk.meshes.push_back(meshData[i].asString());
+        for(size_t i = 0; i < meshData.size(); i++) {
+            chunk.meshes.push_back(meshData[static_cast<int>(i)].asString());
             // LOG_INFO(meshData[i].asString());
         }
